+++ app/src/main/java/com/android/sample/ui/map/MapScreen.kt
@@ -1,9 +1,7 @@
 package com.android.sample.ui.map
 
 import android.Manifest
-import android.content.Context
 import android.content.pm.PackageManager
-import android.graphics.Bitmap
 import android.util.Log
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
@@ -16,10 +14,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -32,35 +26,18 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
-import com.android.sample.R
 import com.android.sample.ui.navigation.BottomNavigationMenu
 import com.android.sample.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.android.sample.ui.navigation.NavigationActions
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.location.Priority.PRIORITY_BALANCED_POWER_ACCURACY
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.model.BitmapDescriptor
-import com.google.android.gms.maps.model.BitmapDescriptorFactory
-import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.tasks.CancellationTokenSource
-import com.google.maps.android.compose.CameraPositionState
-import com.google.maps.android.compose.Circle
-import com.google.maps.android.compose.GoogleMap
-import com.google.maps.android.compose.Marker
-import com.google.maps.android.compose.MarkerState
-import com.google.maps.android.compose.rememberCameraPositionState
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
@@ -184,126 +161,4 @@ fun MapScreen(navigationActions: NavigationActions, currentMusicPlayed: String? 
       }
     }
   }
-}
-
-@Composable
-fun GoogleMapView(
-    currentPosition: MutableState<LatLng>,
-    moveToCurrentLocation: MutableState<CameraAction>,
-    modifier: Modifier,
-    locationPermitted: Boolean,
-    radius: Double = 1000.0
-) {
-  val coroutineScope = rememberCoroutineScope()
-
-  val mapVisible by remember { mutableStateOf(true) }
-
-  val cameraPositionState = rememberCameraPositionState()
-
-  LaunchedEffect(moveToCurrentLocation.value, Unit) {
-    if (moveToCurrentLocation.value == CameraAction.MOVE) {
-      coroutineScope.launch {
-        cameraPositionState.move(
-            update =
-                CameraUpdateFactory.newCameraPosition(
-                    CameraPosition.fromLatLngZoom(currentPosition.value, defaultZoom)))
-        moveToCurrentLocation.value = CameraAction.NO_ACTION
-      }
-    } else if (moveToCurrentLocation.value == CameraAction.ANIMATE) {
-      coroutineScope.launch {
-        cameraPositionState.animate(
-            update =
-                CameraUpdateFactory.newCameraPosition(
-                    CameraPosition.fromLatLngZoom(currentPosition.value, defaultZoom)),
-            durationMs = 1000)
-        moveToCurrentLocation.value = CameraAction.NO_ACTION
-      }
-    }
-  }
-
-  val context = LocalContext.current
-
-  if (mapVisible) {
-    Box(Modifier.fillMaxSize()) {
-      GoogleMap(modifier = modifier, cameraPositionState = cameraPositionState) {
-        currentPosition.value.let { location ->
-          Marker(
-              state = MarkerState(position = location),
-              title = "You are here",
-              icon = getBitmapDescriptorFromDrawableResource(R.drawable.me_position, context),
-              anchor = Offset(0.5f, 0.5f))
-          Circle(
-              center = location,
-              radius = radius,
-              strokeColor = Color(0xFF5F2A83),
-              strokeWidth = 3f,
-              fillColor = Color(0x215F2A83))
-        }
-      }
-      Box(modifier = Modifier.align(Alignment.TopEnd).testTag("currentLocationButton")) {
-        CurrentLocationCenterButton(currentPosition.value, cameraPositionState)
-      }
-    }
-  }
-}
-
-fun getBitmapDescriptorFromDrawableResource(resourceId: Int, context: Context): BitmapDescriptor {
-  val drawable = ContextCompat.getDrawable(context, resourceId)
-  val canvas = android.graphics.Canvas()
-  val bitmap =
-      Bitmap.createBitmap(
-          drawable!!.intrinsicWidth, drawable.intrinsicHeight, Bitmap.Config.ARGB_8888)
-  canvas.setBitmap(bitmap)
-  drawable.setBounds(0, 0, canvas.width, canvas.height)
-  drawable.draw(canvas)
-  return BitmapDescriptorFactory.fromBitmap(bitmap)
-}
-
-@Composable
-fun PlayerCurrentMusicItem(musique: String?) {
-  if (musique != null) {
-    Text(
-        modifier =
-            Modifier.fillMaxWidth().height(24.dp).padding(horizontal = 16.dp).testTag("playerText"),
-        text = "listening : $musique",
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight(400),
-        color = Color(0xFF5F2A83),
-        textAlign = TextAlign.Center)
-  } else {
-    Text(
-        modifier =
-            Modifier.fillMaxWidth().height(24.dp).padding(horizontal = 16.dp).testTag("playerText"),
-        text = "not listening yet",
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight(400),
-        color = Color(0xFF5F2A83),
-        textAlign = TextAlign.Center)
-  }
-}
-
-@Composable
-fun CurrentLocationCenterButton(
-    currentLocation: LatLng?,
-    cameraPositionState: CameraPositionState
-) {
-  FloatingActionButton(
-      containerColor = Color.White,
-      onClick = {
-        currentLocation?.let { location ->
-          cameraPositionState.move(CameraUpdateFactory.newLatLngZoom(location, defaultZoom))
-        }
-      },
-      modifier =
-          Modifier.padding(16.dp)
-              .size(48.dp)
-              .shadow(elevation = 8.dp, shape = RoundedCornerShape(8.dp), clip = false)
-              .background(color = Color.White, shape = RoundedCornerShape(8.dp))
-              .testTag("currentLocationFab")) {
-        Icon(
-            painter = painterResource(id = R.drawable.location_arrow_1),
-            contentDescription = "Current Position",
-            modifier = Modifier.size(30.dp).background(Color.White).testTag("currentLocationIcon"),
-            tint = Color.Unspecified)
-      }
 }
